host on heroku, update update diaries page
