+++ diary/src/App.js
@@ -3,7 +3,7 @@ import { Routes, Route } from "react-router-dom";
 
 import NavMenu from "./components/partials/navMenu.js";
 import HomePage from "./components/homePage.js";
-import DiaryPage from "./components/diariesPage.js";
+import DiaryPage from "./components/diaryPage/diariesPage.js";
 import EntriesPage from "./components/entriesPage.js";
 
 function App() {
