+++ diary/src/components/diaryPage/diariesPage.js
@@ -1,17 +1,18 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import DiaryService from "../service/diaryService.js";
-import { DiaryModal } from "./modalPopup.js";
-import diaryImg from "../assets/diary.jpg";
-import "../styles/DiaryStyles/DiaryPage.css";
-import "../styles/DiaryStyles/DiaryModal.css";
+import DiaryService from "../../service/diaryService.js";
+import DiaryModal from "./diaryModal.js";
+import diaryImg from "../../assets/diary.jpg";
+import "../../styles/DiaryStyles/DiaryPage.css";
 
 // Button component opens Modal Form to create new Diary
-const CreateDiaryButton = ({ openModalHandler }) => {
+const CreateDiaryButton = ({ createBtnHandler }) => {
   return (
-    <button onClick={openModalHandler} type="button">
-      Create a Diary
-    </button>
+    <div className="modal-popup-btn-container">
+      <button className="diary-modal-popup-btn" onClick={createBtnHandler} type="button">
+        Create a Diary
+      </button>
+    </div>
   );
 };
 
@@ -51,47 +52,11 @@ const DiaryList = ({ diaries }) => {
 const DiaryPage = () => {
   const [diaries, setDiaries] = useState([]);
   const [isVisibleModal, setVisibleModal] = useState(false);
-  const [diaryName, setDiaryName] = useState("");
-  const [diaryDescription, setDiaryDescription] = useState("");
-
-  const nameHandler = (event) => {
-    event.preventDefault();
-    setDiaryName(event.target.value);
-    console.log(event.target.value);
-  };
-
-  const descriptionHandler = (event) => {
-    event.preventDefault();
-    setDiaryDescription(event.target.value);
-    console.log(event.target.value);
-  };
 
   // event handler for "Create a Diary" button
-  const modalDiaryHandler = (event) => {
-    event.preventDefault();
+  const createButtonHandler = (event) => {
     setVisibleModal(true);
     console.log("popup");
-  };
-
-  // event handler Diary Form submission
-  const submitDiaryFormHandler = (event) => {
-    event.preventDefault();
-    const newDiary = {
-      diary_name: diaryName,
-    };
-
-    DiaryService.createDiary(newDiary)
-      .then((res) => {
-        setVisibleModal(false);
-        console.log("success");
-      })
-      .catch((err) => {
-        console.log("was not able to create diary");
-        console.log(err);
-      });
-
-    // setDiaryName("");
-    // setDiaryDescription("");
   };
 
   // fetch all diaries from database - no page reload
@@ -107,16 +72,12 @@ const DiaryPage = () => {
   return (
     <div className="diary-hub">
       <h1 className="diary-hub-heading">Diaries</h1>
-      <CreateDiaryButton openModalHandler={modalDiaryHandler} />
+      {isVisibleModal ? (
+        <DiaryModal />
+      ) : (
+        <CreateDiaryButton createBtnHandler={createButtonHandler} />
+      )}
       <DiaryList diaries={diaries} />
-      <DiaryModal
-        formHandler={submitDiaryFormHandler}
-        isVisible={isVisibleModal}
-        name={diaryName}
-        nameHandler={nameHandler}
-        description={diaryDescription}
-        descriptionHandler={descriptionHandler}
-      />
     </div>
   );
 };
