+++ diary/src/utils/formattedDate.js
@@ -1,19 +1,19 @@
 const getDateInMmDdYyyy = function () {
-    let date = new Date();
-    let month = date.getMonth() + 1;
-    let day = date.getDate();
-    let year = date.getFullYear();
+  let date = new Date();
+  let month = date.getMonth() + 1;
+  let day = date.getDate();
+  let year = date.getFullYear();
 
-    if (day < 9) {
-        day = "0" + day;
-    }
+  if (day < 9) {
+    day = "0" + day;
+  }
 
-    if (month < 10) {
-        month = "0" + month;
-    }
+  if (month < 10) {
+    month = "0" + month;
+  }
 
-    const fullDate = month + "/" + day + "/" + year;
-    return fullDate;
+  const fullDate = month + "/" + day + "/" + year;
+  return fullDate;
 };
- 
-module.exports = getDateInMmDdYyyy;
+
+export default getDateInMmDdYyyy;
